chore(backend): scaffold personal best calculations
